Move initial state option out of WeakSubjectivityOptions and into Eth2NetworkOptions.
